Refactor: pinned version restrictions
Github alerts prompted this.

On Ruby 3.0.1 I could not do a `bundle install`, so I have tried to pin
some dependencies in the most permissive way possible.

The keyword args in Ruby 3 don't play well with Ruby 2.1 early keyword args.
Minitest 5.x has upgrade complaints as well.

Bundler was relaxed to 1.x or 2.x (from 1.x) but CVEs suggest it's
better to bump it to v2.2.10 at the lowest.